backend/requirements.txt
# ─────────────────────────────────────────────
# UPSC Intelligence System — Backend Dependencies
# ─────────────────────────────────────────────

# Web Framework
fastapi==0.111.0
uvicorn[standard]==0.30.1

# Database
sqlalchemy==2.0.30
asyncpg==0.29.0          # async PostgreSQL driver
psycopg2-binary==2.9.9   # sync driver (for migrations)
alembic==1.13.1          # DB migrations

# Vector Search
pgvector==0.2.5

# PDF Processing
pymupdf==1.24.3          # text + image extraction from PDFs
ocrmypdf==16.3.0         # OCR for scanned PDFs
pytesseract==0.3.10      # Tesseract OCR wrapper
Pillow==10.3.0           # image processing

# LLM & Embeddings
openai==1.30.1           # GPT-4o for captioning + tagging + answers
tiktoken==0.7.0          # token counting

# Data Validation
pydantic==2.7.1
pydantic-settings==2.3.0

# Background Tasks
celery==5.4.0            # async task queue (OCR, embedding jobs)
redis==5.0.4             # message broker for Celery

# Utilities
python-multipart==0.0.9  # file uploads in FastAPI
python-dotenv==1.0.1     # .env config loading
httpx==0.27.0            # async HTTP client
numpy==1.26.4            # vector math
tenacity==8.3.0          # retry logic for LLM calls
loguru==0.7.2            # structured logging

# Dev / Testing
pytest==8.2.2
pytest-asyncio==0.23.7
aiosqlite==0.20.0        # async SQLite for test database

# Security / Rate Limiting
slowapi==0.1.9           # API throttling per IP
